Add temporary workaround for remaining timescale issue
This should be the last of the causes of the timescale issues (the other ones
being blocking IO in the game loop, and window message loop issues).

This one is strongly related to the log line "err:system:update_display_cache
Failed to read display config". It is likely closely related to similar issues
with xvfb on newer wine versions.

This bug was difficult to track down since it doesn't cause hangs, thrashing, or
unexpected delays in any single process, but something about the way r2, wine,
and the display server (x11 or wayland) interact causes this to trigger the
timescale issue. There is also a race condition somewhere, making this harder to
reproduce on slower hardware.

The bug can be worked around by using the wine virtual desktop config to avoid
querying the display server for physical display layout.

A more permanent fix would involve fully emulating physical displays in the
display server, or implementing a more complete dummy wine graphics driver based
on nulldrv (the issues with the previous attempt are likely closely related to
this).

Many thanks to @wolf109909 for tracking this down and assisting with testing
various fixes and workarounds..

Co-Authored-By: wolf109909 <84360921+wolf109909@users.noreply.github.com>